backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
google-generativeai
pymongo[srv]
langchain-mongodb
psycopg2-binary
sqlalchemy
pymupdf
pandas
openpyxl
pydantic
pydantic-settings
python-dotenv
python-multipart
